mental_health_quiz/requirements.txt
mysql-connector-python
matplotlib